// front/src/api/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Upload: any;
};

export type Discussion = {
  __typename?: 'Discussion';
  id: Scalars['ID'];
  name: Scalars['String'];
};

export enum FhStatus {
  Friend = 'FRIEND',
  Suggestion = 'SUGGESTION',
  RequestFromMe = 'REQUEST_FROM_ME',
  RequestToMe = 'REQUEST_TO_ME'
}

export type Message = {
  __typename?: 'Message';
  id: Scalars['ID'];
  content: Scalars['String'];
  createdAt: Scalars['String'];
  discussionId: Scalars['String'];
  senderId: Scalars['String'];
  sender?: Maybe<User>;
  views?: Maybe<Array<User>>;
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['ID'];
  username: Scalars['String'];
  sex: Scalars['Boolean'];
  birthday: Scalars['String'];
  visible: Scalars['Boolean'];
  connected: Scalars['Boolean'];
  photo?: Maybe<Scalars['String']>;
  photoPublicId?: Maybe<Scalars['String']>;
  fhStatus?: Maybe<FhStatus>;
};